+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.PlaylistMaker"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.PlayerActivity"
+            android:name=".presentation.audioplayer.PlayerActivity"
             android:exported="false" />
         <activity
             android:name=".SearchActivity"
