+++ app/src/main/java/com/example/playlistmaker/SearchAdapter.kt
@@ -12,7 +12,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.example.playlistmaker.presentation.PlayerActivity
+import com.example.playlistmaker.domain.Track
+import com.example.playlistmaker.presentation.audioplayer.PlayerActivity
 import java.util.ArrayList
 
 class SearchAdapter(diplayedList: ArrayList<Track>, activity: SearchActivity) :
