+++ app/src/main/java/com/example/playlistmaker/domain/Tracks.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
