+++ app/src/main/java/com/example/playlistmaker/domain/UnparsedTrack.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.domain
+
+class unparsedTrack(
+    val trackName: String,
+    val artistName: String,
+    val trackTimeMillis: Long,
+    val artworkUrl100: String,
+    val trackId : Long,
+    val collectionName : String?,
+    val releaseDate : String,
+    val primaryGenreName : String,
+    val country : String,
+    val previewUrl : String
+)
