+++ app/src/main/java/com/example/playlistmaker/domain/searchServerResponse.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.domain
+
+data class searchServerResonse(
+    val resultCount:Int,
+    val results: List<unparsedTrack>
+)
+{
+}
