+++ app/src/main/java/com/example/playlistmaker/presentation/audioplayer/PlayerActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.presentation
+package com.example.playlistmaker.presentation.audioplayer
 
 
 import android.annotation.SuppressLint
@@ -20,7 +20,7 @@ import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.Constants
 import com.example.playlistmaker.MediaPlayer
 import com.example.playlistmaker.R
-import com.example.playlistmaker.Track
+import com.example.playlistmaker.domain.Track
 import java.util.Locale
